backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
pydantic-settings==2.0.3
yfinance==0.2.31